prepare-and-start.sh: First create current group and then docker
We create two groups right now: one for docker and one for the
user running the script (to avoid any permission issues). The docker group
is created without specifying the GID so it will take the first available one
(in container that is 1000). The user GID can be anything from 1000 up.
If the user has GID 1000 and we first create docker group (this will take 1000)
and then the user group (1000), we will get an error. Avoid this by creating
the current user group first and then the docker one.

Signed-off-by: Andrei Gherzan <andrei@resin.io>