Dockerfile_yocto-build-env: Update docker base image to ubuntu 16.04 (LTS)
Also update nodejs to v8 and remove npm from being installed as the nodejs package
now contains the npm binary and it will conflict when "apt get install"-ing npm.

We also need to install the "locales" package as it has been removed from the ubuntu 16.04
base image in order to reduce the base image size.

Signed-off-by: Florin Sarbu <florin@resin.io>