jenkins_build.sh: Switch s3cmd with s4cmd
The resin-img docker image has changed to open-balena-base 8.0.0
which in turn is based on debian:buster. This version of debian
does not include the s3cmd package in it's repositories anymore.
Instead it provides a replacement package called s4cmd
(https://packages.debian.org/buster/s4cmd) which "is intended as
an alternative to the existing s3cmd tool, enhancing performance
and support for large files, and coming with a number of additional
features and fixes".

Change-type: minor
Signed-off-by: Florin Sarbu <florin@balena.io>