jenkins_build.sh: Use canned ACL and remove obsolete option
Since s4cmd does not have the "--acl-public" option we use the new
--API-ACL option that allows us to set canned ACLs for the objects
we upload to S3. We use the "public-read" canned ACL that does:
"Owner gets FULL_CONTROL. The AllUsers group gets READ access."
Also we remove the "--skip-existing" option which is not present
in s4cmd.

Change-type: patch
Signed-off-by: Florin Sarbu <florin@balena.io>