jenkins_build.sh: Use dsync instead of deprecated sync command
The sync command is deprecated in s4cmd in favour of dsync. And also
the existing deprecated sync command in s4cmd does the sync in a
different manner from the s3cmd sync command in the sense that it
makes a copy of the local dir in the remote dir and in our case we
end up with duplicate dirs in the target path.

Change-type: patch
Signed-off-by: Florin Sarbu <florin@balena.io>