build-device-type-json: show stderr on failure
When building a device type JSON manifest, any syntax error is handled
by printing a message about node being unavailable, even when it's
installed and working.

Check that node is installed and executable before building the
manifest, and exit early with the existing message if not. Otherwise,
print stderr to the console for debugging.

Change-type: patch
Signed-off-by: Joseph Kogut <joseph@balena.io>