jenkins_build.sh: publish explicitly named balena-raw and balena-flasher images
At this moment we are providing a single balena.img file that is either
raw or flasher based on the coffee script. We want to publish both
for some device types, therefore split the single file into balena-raw.img
and balena-flasher.img.

Temporarily keep the existing balena.img for backwards compatibility until
all coffee scripts are updated and image maker is aware of the split.

Change-type: minor
Signed-off-by: Michal Toman <michalt@balena.io>