balena-lib: Do not err when sourcing without a repository
A global variable uses git to retrieve a sha1 revision. If this fails
with the errexit option the script exists after sourcing this file.

This commit masks the error as it may not be needed in the sourcing
script.

Also, fix the length of the short git release to avoid different git
clients configuration mismatches.

Change-type: patch
Signed-off-by: Alex Gonzalez <alexg@balena.io>