Generate config.js from test suites in meta-balena
For each test suite in meta-balena we need a a blob
of javascript added to the config.js manifest for leviathan.

Also allow overriding the default config values with a
suite.env file in the root of the suite directory.

Change-type: patch
Changelog-entry: Generate config.js from test suites in meta-balena
Signed-off-by: Kyle Harding <kyle@balena.io>