balena-docker: Allow to control iptables and ipmasq flags
When running multiple daemons, we don't want them to clash managing
iptables so we start redundant daemons with iptables and ipmasq set to
false.

By default we enable both which is the current dockerd default.

When running a single docker daemon without the iptables and ipmasq flags,
containers on the default bridge network cannot communite with the
outside as no iptables rules are set.

Change-type: patch
Signed-off-by: Alex Gonzalez <alexg@balena.io>