versionbot: Add changelog yml file
This file allows other components to uniquely parse the information that
is contained in the changelog. It will be automatically managed by
versionist by appending the new commits on top. This is needed to
provide nested-changelogs.

Generated with product-os/versionist/blob/master/scripts/generate-changelog.sh

Change-type: patch
Changelog-entry: Add a parsable representation of the changelog
Signed-off-by: Alex Gonzalez <alexg@balena.io>