jenkins_build-blocks: Separate recipes and packages
Currently we are reading a package list from the contracts that mixes
recipe and package names, and we run bitbake with the `-k` flag not to
stop the build when a package name fails to match a recipe.

The contracts have now been modified to specify either a `sw.recipe.yocto`
or a `sw.package.yocto` so we can distinguish between them and use
recipe names to build and package names to install.

We also remove the continue flag from bitbake.

Change-type: patch
Signed-off-by: Alex Gonzalez <alexg@balena.io>