balena-lib: improve base tag detection
When a ESR release is deployed a tag with the base meta-balena version
is created. This is used by the API to check for a valid OS version
for updates.

The current mechanism to find the base version only provides an ESR
version for the first commit after the branch has been created.

Using merge-base to find the common ancestor and `tag --points-at` to
find the actual tag works for all commits after the branch is created.

Change-type: patch
Signed-off-by: Alex Gonzalez <alexg@balena.io>