Simplify matrix generation
We previously had a few inputs that were expected to be comma-separated
lists (like 'a,b,c') and later steps would then convert them to a JSON
array ('["a", "b", "c"]') so that they could be used to generate the
matrix strategy.

With this commit, we require those inputs to be formatted as JSON arrays
from the start. This allows us to remove the conversion steps, thus
simplifying the matrix generation. It also makes the workflow as a whole
a bit more consistent, given that we already have other inputs that are
expected to be JSON arrays.

Signed-off-by: Leandro Motta Barros <leandro@balena.io>
Change-type: patch