Simplify check for secure boot
We were previously introducing a variable that was essentially
replicating the contents of an input. This commit makes use of the input
directly.

Incidentally, the previous code also had a small bug/typo, in which we
mixed a test for string length with a test for string equality.

Signed-off-by: Leandro Motta Barros <leandro@balena.io>
Change-type: patch