Show unexpected input logic so we can still write safe code
We do not expect force-finalize to be true on any event types
except workflow dispatch, but we need to handle it correctly
either way.

Similarly, we treat auto-finalize as no-op for events
other than tag push events, but we should still show the logic
gates and how the various values entangle.

Signed-off-by: Kyle Harding <kyle@balena.io>