Add unique unique identifiers to summary job
The "All jobs" job is used for branch protection
as it captures the overall success/fail of all
the test matrices and build.

However without a unique name we are unable
to identify the device type in order to mark it as required.

This change adds a single instance matrix
to the build job and as a result will append the machine
name and deploy environment in the job name.

Change-type: patch
Signed-off-by: Kyle Harding <kyle@balena.io>