Append to MIRRORS instead of PREMIRRORS to limit egress
For the first deployment of the S3 download mirror we will
err on the side of cost savings, and use the mirror as a
fallback not the primary source of downloads.

Revert this if we want the benefits of using our own mirrors
as the primary source, like download speed and no rate limits.

Change-type: patch
Signed-off-by: Kyle Harding <kyle@balena.io>