Separate build jobs by environments
This allows more granular control over which environments are used
for which steps, and avoids requiring a single environment to define
how to build, sign, and publish.

This is better for security as it allows each job to only have
the secrets it needs, and not all secrets for all steps.

We retain backwards compatibility by falling back to the original
monolith environment input for now.

Change-type: minor
Signed-off-by: Kyle Harding <kyle@balena.io>