barys: allow for custom templates
By default barys assumes there is only one integration layers and looks
for the custom sample templates in there. This might not be the case
if there are feature layers like adding secure boot support, so this
commit allows to specify which layer to look the sample templates from.

Change-type: patch
Signed-off-by: Alex Gonzalez <alexg@balena.io>