Dedupe raw or flasher images files during artifact prepare
Currently generic device types are uploading 3 images in the
test artifacts

balena.img.enc
balena-flasher.img.enc
balena-raw.img.enc

But balena.img is always going to match either the flasher or
the raw image depending on the device spec.

This change will replace balena.img(.enc) with a symlink when
possible in only the test artifacts to reduce the upload and download
size.

Change-type: patch
Signed-off-by: Kyle Harding <kyle@balena.io>