balena-lib: Keep matching the existing kontron-come-xelx contract slug with the older device-type.json
For the kontron-come-xelx device type we received a BSP update on which the hw vendor changed
the Yocto machine name so we were forced to adapt to this by changing the relevant .coffee file
in the device repo. This made it that the generated device-type.json is now different for this
device type's contract slug so we need to make sure we keep using the correct correspondence.

Change-type: patch
Signed-off-by: Florin Sarbu <florin@balena.io>